cloud-functions/table-manager/requirements.txt
google-cloud-bigquery==3.25.0
google-cloud-pubsub==2.23.0
functions-framework==3.8.1 